unify problem.py and generate.py
